I'm calling lights done. FPS are decent (however, I can easily see myself optimizing the lights further in the future). Moving on...

+++ trunk/src/com/kobaj/openglgraphics/BaseLightShader.java
@@ -23,7 +23,7 @@ public abstract class BaseLightShader
 	public int my_brightness_handle;
 	
 	//effects
-	public float my_brightness;
+	public double my_brightness;
 	public FloatBuffer my_color;
 	
 	public BaseLightShader()
